refactor: 클래스를 domain, view 패키지에 집어넣음
