+++ src/main/java/racingcar/domain/Cars.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.domain;
+
+import racingcar.view.OutputView;
 
 import java.util.ArrayList;
 import java.util.List;
